+++ projects/commonlibs/zocromas_qstd/src/qstd_mstmt_props.c
@@ -63,30 +63,26 @@ mas_qstd_mstmt_insget_props_id( mas_qstd_t * qstd, unsigned long long data_id, c
   {
     mysqlpfs_mstmt_t *mstmt = mas_qstd_mstmt_get( qstd, STD_MSTMT_INSERT_PROPS );
 
-  /* const schar *s = "INSERT  INTO fileprops(data_id,detype,mode,uid,gid,atim,mtim,ctim,size,rdev,blksize,blocks) " */
-  /*         "VALUES (?,?,?,?,?,FROM_UNIXTIME(?),FROM_UNIXTIME(?),FROM_UNIXTIME(?),?,?,?,?)";                        */
-
-/* data_id,detype,mode,uid,gid,atim,mtim,ctim,size,rdev,blksize,blocks */
     QRGP( mstmt );
     rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 0, data_id, FALSE ) );
     rC( mas_mysqlpfs_mstmt_set_param_string( mstmt, 1, sdetype ) );
 
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 2, stat->st_mode, stat->st_mode ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 3, stat->st_uid, stat->st_uid ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 4, stat->st_gid, stat->st_gid ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 2, ( unsigned long long ) stat->st_mode, stat->st_mode ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 3, ( unsigned long long ) stat->st_uid, stat->st_uid ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 4, ( unsigned long long ) stat->st_gid, stat->st_gid ? FALSE : TRUE ) );
   /*
      struct timespec {
      time_t tv_sec;
      long tv_nsec;
      };
    */
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 5, stat->st_atim.tv_sec, stat->st_atim.tv_sec ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 6, stat->st_mtim.tv_sec, stat->st_mtim.tv_sec ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 7, stat->st_ctim.tv_sec, stat->st_ctim.tv_sec ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 8, stat->st_size, FALSE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 9, stat->st_rdev, stat->st_rdev ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 10, stat->st_blksize, stat->st_blksize ? FALSE : TRUE ) );
-    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 11, stat->st_blocks, stat->st_blocks ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 5, ( unsigned long long ) stat->st_atim.tv_sec, stat->st_atim.tv_sec ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 6, ( unsigned long long ) stat->st_mtim.tv_sec, stat->st_mtim.tv_sec ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 7, ( unsigned long long ) stat->st_ctim.tv_sec, stat->st_ctim.tv_sec ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 8, ( unsigned long long ) stat->st_size, FALSE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 9, ( unsigned long long ) stat->st_rdev, stat->st_rdev ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 10, ( unsigned long long ) stat->st_blksize, stat->st_blksize ? FALSE : TRUE ) );
+    rC( mas_mysqlpfs_mstmt_set_param_longlong( mstmt, 11, ( unsigned long long ) stat->st_blocks, stat->st_blocks ? FALSE : TRUE ) );
 
     rC( mas_mysqlpfs_mstmt_execute( mstmt ) );
